backend/requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
Flask-Bcrypt==1.0.1
Flask-Limiter==3.12
Flask-CORS==5.0.1
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dotenv==1.0.1
gunicorn==23.0.0
